Add a way to force single carriageways and compute them
The computation is based on cases where "normal" lines meet those with a
'psv' attribute, which seem to be abnormal and create issues with dual
carriageways.

In general, now a 'forced_single' attribute can be set to True for every
feature, which will force the script to create single carriageway for a
road no matter what its attributes/width are.